requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
spacy==3.7.5
pdfminer.six==20231228
chardet==5.2.0
scikit-learn==1.5.1